application/chat-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pillow==10.2.0
pyyaml==6.0.1
torch==2.2.1
torchvision==0.17.1
numpy<2
timm>=0.9.8
